Revert int to long conversation

+++ platforms/extensibility/plugin-development/src/main/java/org/gradle/plugin/devel/tasks/internal/ValidationProblemSerialization.java
@@ -25,13 +25,6 @@ import com.google.gson.reflect.TypeToken;
 import com.google.gson.stream.JsonReader;
 import com.google.gson.stream.JsonToken;
 import com.google.gson.stream.JsonWriter;
-import org.gradle.api.problems.internal.DefaultFileLocation;
-import org.gradle.api.problems.internal.DefaultLineInFileLocation;
-import org.gradle.api.problems.internal.DefaultOffsetInFileLocation;
-import org.gradle.api.problems.internal.DefaultPluginIdLocation;
-import org.gradle.api.problems.internal.DefaultProblem;
-import org.gradle.api.problems.internal.DefaultProblemCategory;
-import org.gradle.api.problems.internal.DefaultTaskPathLocation;
 import org.gradle.api.problems.internal.DocLink;
 import org.gradle.api.problems.internal.FileLocation;
 import org.gradle.api.problems.internal.LineInFileLocation;
@@ -39,6 +32,13 @@ import org.gradle.api.problems.internal.OffsetInFileLocation;
 import org.gradle.api.problems.internal.Problem;
 import org.gradle.api.problems.internal.ProblemCategory;
 import org.gradle.api.problems.internal.ProblemLocation;
+import org.gradle.api.problems.internal.DefaultFileLocation;
+import org.gradle.api.problems.internal.DefaultLineInFileLocation;
+import org.gradle.api.problems.internal.DefaultOffsetInFileLocation;
+import org.gradle.api.problems.internal.DefaultPluginIdLocation;
+import org.gradle.api.problems.internal.DefaultProblem;
+import org.gradle.api.problems.internal.DefaultProblemCategory;
+import org.gradle.api.problems.internal.DefaultTaskPathLocation;
 import org.gradle.internal.reflect.validation.TypeValidationProblemRenderer;
 
 import javax.annotation.Nonnull;
@@ -243,10 +243,10 @@ public class ValidationProblemSerialization {
         private static FileLocation readObject(JsonReader in) throws IOException {
             String subtype = null;
             String path = null;
-            Long offset = null;
-            Long line = null;
-            Long column = null;
-            Long length = null;
+            Integer offset = null;
+            Integer line = null;
+            Integer column = null;
+            Integer length = null;
             while (in.hasNext()) {
                 String name = in.nextName();
                 switch (name) {
@@ -259,19 +259,19 @@ public class ValidationProblemSerialization {
                         break;
                     }
                     case "offset": {
-                        offset = in.nextLong();
+                        offset = in.nextInt();
                         break;
                     }
                     case "line": {
-                        line = in.nextLong();
+                        line = in.nextInt();
                         break;
                     }
                     case "column": {
-                        column = in.nextLong();
+                        column = in.nextInt();
                         break;
                     }
                     case "length": {
-                        length = in.nextLong();
+                        length = in.nextInt();
                         break;
                     }
                     default:
